+++ frontend/src/metabase/query_builder/components/view/QuestionDisplayToggle/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./QuestionDisplayToggle";
