--- frontend/src/metabase/query_builder/components/view/QuestionTableToggle/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./QuestionTableToggle";
